重新组织skill代码组织结构
