+++ apps/electron/src/renderer/components/app-shell/AppShell.tsx
@@ -86,8 +86,8 @@ import {
 import type { SettingsSubpage } from "../../../shared/types"
 import { SourcesListPanel } from "./SourcesListPanel"
 import { ArchivedSessionsPanel } from "./ArchivedSessionsPanel"
-import { SkillsListPanel } from "./SkillsListPanel"
-import { SkillImportModal } from "./SkillImportModal"
+import { SkillsListPanel } from "./skill/SkillsListPanel"
+import { SkillImportModal } from "./skill/SkillImportModal"
 import { AutomationsListPanel } from "../automations/AutomationsListPanel"
 import { APP_EVENTS, AGENT_EVENTS, type AutomationFilterKind, AUTOMATION_TYPE_TO_FILTER_KIND } from "../automations/types"
 import { useAutomations } from "@/hooks/useAutomations"
@@ -115,7 +115,7 @@ import {
   createSkillsSidebarItems,
   LOCAL_SKILLS_NAV_ID,
   SKILL_MARKETPLACE_NAV_ID,
-} from "./skills-sidebar-items"
+} from "./skill/skills-sidebar-items"
 import {
   loadRightSidebarOpenPreference,
   persistRightSidebarOpenPreference,
