+++ apps/electron/src/renderer/components/app-shell/MainContentPanel.tsx
@@ -45,7 +45,7 @@ import { AutomationInfoPage } from '../automations/AutomationInfoPage'
 import type { ExecutionEntry } from '../automations/types'
 import { automationsAtom } from '@/atoms/automations'
 import { SendResourceToWorkspaceDialog, type SendResourceType } from './SendResourceToWorkspaceDialog'
-import { SkillMarketplacePage } from './SkillMarketplacePage'
+import { SkillMarketplacePage } from './skill/SkillMarketplacePage'
 
 export interface MainContentPanelProps {
   /** Whether both sidebar and navigator are hidden (focus mode / CMD+.) */
