+++ apps/electron/src/renderer/components/app-shell/skill/__tests__/skills-sidebar-items.test.ts
