+++ apps/electron/src/renderer/components/app-shell/skill/mock-data.ts
@@ -1,0 +1,411 @@
+import type { CopawMarketSkill } from '@craft-agent/shared/skills'
+import type { MarketplaceSkillListing } from './types'
+import type { LoadedSkill } from '../../../../shared/types'
+
+// ============================================================================
+// Mock switch — set to false to use real API
+// ============================================================================
+
+export const USE_MOCK_MARKET = true
+
+export const MOCK_MARKET_SKILLS: CopawMarketSkill[] = [
+  {
+    fileKey: 'mock-key-1',
+    userName: '张三',
+    employeeId: 'MOCK_CURRENT_USER',
+    department: '研发部',
+    name: 'code-review',
+    chineseName: '代码审查助手',
+    description: '自动审查代码质量，给出改进建议，支持多种编程语言。',
+    tag: 'A',
+    hot: 128,
+    createdAt: '2024-01-15T08:00:00Z',
+    version: '20240115080000',
+  },
+  {
+    fileKey: 'mock-key-2',
+    userName: '李四',
+    employeeId: 'MOCK_CURRENT_USER',
+    department: 'DevOps团队',
+    name: 'k8s-deploy',
+    chineseName: 'K8s 部署助手',
+    description: '一键生成 Kubernetes 部署配置，支持 Deployment、Service、Ingress 等资源。',
+    tag: 'B',
+    hot: 256,
+    createdAt: '2024-02-20T10:00:00Z',
+    version: '20240220100000',
+  },
+  {
+    fileKey: 'mock-key-3',
+    userName: '王五',
+    employeeId: 'EMP003',
+    department: '测试部',
+    name: 'test-generator',
+    chineseName: '单测生成器',
+    description: '根据函数签名和逻辑自动生成单元测试用例，覆盖边界条件。',
+    tag: 'A',
+    hot: 87,
+    createdAt: '2024-03-10T14:00:00Z',
+    version: '20240310140000',
+  },
+  {
+    fileKey: 'mock-key-4',
+    userName: '赵六',
+    employeeId: 'EMP004',
+    department: 'DevOps团队',
+    name: 'ci-pipeline',
+    chineseName: 'CI 流水线助手',
+    description: '快速生成 Jenkins / GitLab CI / GitHub Actions 流水线配置文件。',
+    tag: 'B',
+    hot: 312,
+    createdAt: '2024-03-22T09:00:00Z',
+    version: '20240322090000',
+  },
+  {
+    fileKey: 'mock-key-5',
+    userName: '陈七',
+    employeeId: 'EMP005',
+    department: '研发部',
+    name: 'sql-optimizer',
+    chineseName: 'SQL 优化器',
+    description: '分析慢查询，给出索引优化和 SQL 改写建议。',
+    tag: 'A',
+    hot: 45,
+    createdAt: '2024-04-01T11:00:00Z',
+    version: '20240401110000',
+  },
+  {
+    fileKey: 'mock-key-6',
+    userName: '周八',
+    employeeId: 'EMP006',
+    department: 'DevOps团队',
+    name: 'docker-builder',
+    chineseName: 'Docker 镜像构建',
+    description: '智能生成多阶段 Dockerfile，优化镜像体积，支持 Node.js / Python / Java 等常见运行时。',
+    tag: 'B',
+    hot: 198,
+    createdAt: '2024-04-10T09:30:00Z',
+    version: '20240410093000',
+  },
+  {
+    fileKey: 'mock-key-7',
+    userName: '吴九',
+    employeeId: 'EMP007',
+    department: '架构组',
+    name: 'api-designer',
+    chineseName: 'RESTful API 设计师',
+    description: '根据业务描述自动生成 OpenAPI 3.0 规范文档，包含请求体、响应体和错误码定义。',
+    tag: 'A',
+    hot: 163,
+    createdAt: '2024-04-18T15:00:00Z',
+    version: '20240418150000',
+  },
+  {
+    fileKey: 'mock-key-8',
+    userName: '郑十',
+    employeeId: 'EMP008',
+    department: 'DevOps团队',
+    name: 'log-analyzer',
+    chineseName: '日志分析助手',
+    description: '解析应用日志，自动识别异常模式，定位报错根因，生成排查报告。',
+    tag: 'B',
+    hot: 74,
+    createdAt: '2024-05-02T11:00:00Z',
+    version: '20240502110000',
+  },
+  {
+    fileKey: 'mock-key-9',
+    userName: '孙十一',
+    employeeId: 'EMP009',
+    department: '研发部',
+    name: 'commit-writer',
+    chineseName: 'Git Commit 生成器',
+    description: '分析 git diff 内容，自动生成符合 Conventional Commits 规范的提交信息。',
+    tag: 'A',
+    hot: 221,
+    createdAt: '2024-05-15T08:00:00Z',
+    version: '20240515080000',
+  },
+  {
+    fileKey: 'mock-key-10',
+    userName: '李十二',
+    employeeId: 'EMP010',
+    department: '研发部',
+    name: 'doc-writer',
+    chineseName: '技术文档助手',
+    description: '根据代码或功能描述自动撰写技术文档、README 和接口说明，支持中英文输出。',
+    tag: 'A',
+    hot: 309,
+    createdAt: '2024-05-20T14:00:00Z',
+    version: '20240520140000',
+  },
+  {
+    fileKey: 'mock-key-11',
+    userName: '赵十三',
+    employeeId: 'EMP011',
+    department: 'DevOps团队',
+    name: 'infra-cost',
+    chineseName: '云资源成本分析',
+    description: '扫描 Terraform / CloudFormation 配置，估算月度云资源费用并给出优化建议。',
+    tag: 'B',
+    hot: 56,
+    createdAt: '2024-06-01T10:00:00Z',
+    version: '20240601100000',
+  },
+  {
+    fileKey: 'mock-key-12',
+    userName: '钱十四',
+    employeeId: 'EMP012',
+    department: '安全团队',
+    name: 'security-scanner',
+    chineseName: '代码安全扫描',
+    description: '检测代码中的 OWASP Top10 漏洞，包括 SQL 注入、XSS、敏感信息泄露等，输出修复建议。',
+    tag: 'A',
+    hot: 142,
+    createdAt: '2024-06-10T09:00:00Z',
+    version: '20240610090000',
+  },
+]
+
+export const MOCK_SKILLS: MarketplaceSkillListing[] = [
+  {
+    id: 'mkt_docker',
+    slug: 'docker',
+    ownerId: 'owner_mdp',
+    icon: 'D',
+    iconBg: 'bg-blue-500',
+    name: 'Docker',
+    description: '管理容器的构建、运行与生命周期，管理容器的构建、运行与生命周期管理容器的构建、运行与生命周期',
+    owner: 'MDP Labs',
+    category: 'DevOps',
+    tags: ['container', 'devops'],
+    latestVersion: '1.2.0',
+    installCount: 3812,
+    installState: 'install',
+  },
+  {
+    id: 'mkt_kubernetes',
+    slug: 'kubernetes',
+    ownerId: 'owner_mdp',
+    icon: 'K8s',
+    iconBg: 'bg-indigo-500',
+    name: 'Kubernetes',
+    description: '自动化部署、扩缩容与集群管理',
+    owner: 'MDP Labs',
+    category: 'DevOps',
+    tags: ['k8s', 'devops', 'cloud'],
+    latestVersion: '2.0.1',
+    installCount: 2940,
+    installState: 'installed',
+  },
+  {
+    id: 'mkt_github_actions',
+    slug: 'github-actions',
+    ownerId: 'owner_mdp',
+    icon: 'GA',
+    iconBg: 'bg-gray-700',
+    name: 'GitHub Actions',
+    description: '构建和管理 CI/CD 工作流',
+    owner: 'MDP Labs',
+    category: 'DevOps',
+    tags: ['ci', 'cd', 'github'],
+    latestVersion: '1.5.3',
+    installCount: 5120,
+    installState: 'install',
+  },
+  {
+    id: 'mkt_terraform',
+    slug: 'terraform',
+    ownerId: 'owner_mdp',
+    icon: 'TF',
+    iconBg: 'bg-purple-600',
+    name: 'Terraform',
+    description: '基础设施即代码，管理云资源',
+    owner: 'MDP Labs',
+    category: 'DevOps',
+    tags: ['iac', 'cloud', 'devops'],
+    latestVersion: '1.0.0',
+    installCount: 1876,
+    installState: 'install',
+  },
+  {
+    id: 'mkt_prometheus',
+    slug: 'prometheus',
+    ownerId: 'owner_mdp',
+    icon: 'PM',
+    iconBg: 'bg-orange-500',
+    name: 'Prometheus',
+    description: '监控指标采集与告警规则管理',
+    owner: 'MDP Labs',
+    category: 'DevOps',
+    tags: ['monitoring', 'metrics'],
+    latestVersion: '1.1.2',
+    installCount: 1340,
+    installState: 'install',
+  },
+  {
+    id: 'mkt_jenkins',
+    slug: 'jenkins',
+    ownerId: 'owner_mdp',
+    icon: 'JK',
+    iconBg: 'bg-red-500',
+    name: 'Jenkins',
+    description: '配置并触发 Jenkins 构建流水线',
+    owner: 'MDP Labs',
+    category: 'DevOps',
+    tags: ['ci', 'build'],
+    latestVersion: '1.0.4',
+    installCount: 987,
+    installState: 'installed',
+  },
+  {
+    id: 'mkt_doc_gen',
+    slug: 'doc-generator',
+    ownerId: 'owner_community',
+    icon: 'DG',
+    iconBg: 'bg-emerald-500',
+    name: '文档生成器',
+    description: '根据代码自动生成技术文档',
+    owner: '社区',
+    category: '公共',
+    tags: ['docs', 'automation'],
+    latestVersion: '1.3.0',
+    installCount: 2210,
+    installState: 'install',
+  },
+  {
+    id: 'mkt_code_review',
+    slug: 'code-review',
+    ownerId: 'owner_community',
+    icon: 'CR',
+    iconBg: 'bg-cyan-500',
+    name: '代码审查',
+    description: '智能分析代码质量与安全问题',
+    owner: '社区',
+    category: '公共',
+    tags: ['review', 'quality'],
+    latestVersion: '2.1.0',
+    installCount: 3450,
+    installState: 'install',
+  },
+  {
+    id: 'mkt_calendar',
+    slug: 'calendar-assistant',
+    ownerId: 'owner_community',
+    icon: 'CA',
+    iconBg: 'bg-green-500',
+    name: '日历助手',
+    description: '管理日程、会议与提醒事项',
+    owner: '社区',
+    category: '公共',
+    tags: ['calendar', 'productivity'],
+    latestVersion: '1.0.2',
+    installCount: 1560,
+    installState: 'installed',
+  },
+  {
+    id: 'mkt_weekly_report',
+    slug: 'weekly-report',
+    ownerId: 'owner_community',
+    icon: 'WR',
+    iconBg: 'bg-violet-500',
+    name: '周报助手',
+    description: '自动整理工作内容，生成周报草稿',
+    owner: '社区',
+    category: '公共',
+    tags: ['report', 'productivity'],
+    latestVersion: '1.1.0',
+    installCount: 4230,
+    installState: 'install',
+  },
+]
+
+export const MOCK_LOCAL_SKILLS: LoadedSkill[] = [
+  // 市场安装的技能（有 marketplaceOrigin）
+  {
+    slug: 'docker',
+    metadata: { name: 'Docker', description: '管理容器的构建、运行与生命周期' },
+    content: '',
+    path: '/workspace/.agents/docker',
+    source: 'workspace',
+    marketplaceOrigin: {
+      marketplaceId: 'mkt_docker',
+      marketplaceSlug: 'docker',
+      ownerId: 'owner_devops',
+      ownerDisplayName: 'DevOps Team',
+      installedVersion: '1.2.0',
+      installedAt: '2024-01-10T08:00:00Z',
+      lastCheckedAt: '2024-01-10T08:00:00Z',
+      modified: false,
+      sourceBundleHash: 'abc123',
+      safetyStatus: 'ok',
+    },
+  },
+  {
+    slug: 'kubernetes',
+    metadata: { name: 'Kubernetes', description: '自动化部署、扩缩容与集群管理' },
+    content: '',
+    path: '/workspace/.agents/kubernetes',
+    source: 'workspace',
+    marketplaceOrigin: {
+      marketplaceId: 'mkt_kubernetes',
+      marketplaceSlug: 'kubernetes',
+      ownerId: 'owner_devops',
+      ownerDisplayName: 'DevOps Team',
+      installedVersion: '2.0.1',
+      installedAt: '2024-01-12T09:00:00Z',
+      lastCheckedAt: '2024-01-12T09:00:00Z',
+      modified: false,
+      sourceBundleHash: 'def456',
+      safetyStatus: 'ok',
+    },
+  },
+  {
+    slug: 'github-actions',
+    metadata: { name: 'GitHub Actions', description: '构建和管理 CI/CD 工作流' },
+    content: '',
+    path: '/workspace/.agents/github-actions',
+    source: 'workspace',
+    marketplaceOrigin: {
+      marketplaceId: 'mkt_github_actions',
+      marketplaceSlug: 'github-actions',
+      ownerId: 'owner_devops',
+      ownerDisplayName: 'DevOps Team',
+      installedVersion: '1.5.0',
+      installedAt: '2024-01-15T10:00:00Z',
+      lastCheckedAt: '2024-01-15T10:00:00Z',
+      modified: false,
+      sourceBundleHash: 'ghi789',
+      safetyStatus: 'ok',
+    },
+  },
+  // 本地上传的技能（无 marketplaceOrigin）
+  {
+    slug: 'code-review',
+    metadata: { name: '代码审查', description: '智能分析代码质量与安全问题', author: '张三' },
+    content: '',
+    path: '/workspace/.agents/code-review',
+    source: 'workspace',
+  },
+  {
+    slug: 'weekly-report',
+    metadata: { name: '周报助手', description: '自动整理工作内容，生成周报草稿', author: '李四' },
+    content: '',
+    path: '/workspace/.agents/weekly-report',
+    source: 'workspace',
+  },
+  {
+    slug: 'doc-generator',
+    metadata: { name: '文档生成器', description: '根据代码自动生成技术文档', author: '王五' },
+    content: '',
+    path: '/workspace/.agents/doc-generator',
+    source: 'workspace',
+  },
+]
+
+export const HERO_SLIDES = [
+  { icon: 'D', color: 'bg-blue-500', name: 'Docker', prompt: '帮我分析容器内存占用并优化配置' },
+  { icon: 'K8s', color: 'bg-indigo-500', name: 'Kubernetes', prompt: '检查集群状态并扩容 production 命名空间' },
+  { icon: 'GA', color: 'bg-gray-700', name: 'GitHub Actions', prompt: '为我的项目生成一个完整的 CI/CD 流水线' },
+  { icon: 'WR', color: 'bg-violet-500', name: '周报助手', prompt: '整理本周提交记录并生成工作周报' },
+]
