+++ apps/electron/src/renderer/pages/SkillInfoPage.tsx
@@ -12,8 +12,8 @@ import { useEffect, useState, useCallback } from 'react'
 import { Check, X, Minus } from 'lucide-react'
 import { EditPopover, EditButton, getEditConfig } from '@/components/ui/EditPopover'
 import { toast } from 'sonner'
-import { SkillMenu } from '@/components/app-shell/SkillMenu'
-import { LocalSkillMarketplaceStatus } from '@/components/app-shell/SkillMarketplacePage'
+import { SkillMenu } from '@/components/app-shell/skill/SkillMenu'
+import { LocalSkillMarketplaceStatus } from '@/components/app-shell/skill/SkillMarketplacePage'
 import { SkillAvatar } from '@/components/ui/skill-avatar'
 import { routes, navigate } from '@/lib/navigate'
 import { useActiveWorkspace } from '@/context/AppShellContext'
